% Auto-generated: do not edit by hand
\name{''WebsocketClient}

\alias{''WebsocketClient}

\title{WebsocketClient component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
''WebsocketClient(id=NULL, url=NULL, value=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{url}{Character. A url connection that will be used to create the socket.}

\item{value}{Unnamed list. The value displayed in the input.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

